# LabRestApi/src/main/resources/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API для Лабораторной информационной системы",
    "version": "2.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/orders/{id}/status": {
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "Обновить статус заявки",
        "description": "Обновляет поле status у заявки",
        "operationId": "updateOrderStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "REGISTERED",
                "IN_PROGRESS",
                "COMPLETED",
                "CANCELED"
              ]
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Заявка не найдена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Заявка успешно обновлена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Получить все заявки",
        "description": "Возвращает список всех заявок. Реализована пагинация: по умолчанию page=0 & size=50",
        "operationId": "getAllOrders",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Заявки не найдены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Список заявок успешно получен.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Создать новую заявку",
        "description": "Создает новую заявку на основе переданных данных",
        "operationId": "createOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Некорректные данные заявки",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "Заявка успешно создана",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRequestDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders/{orderId}/tests": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Получить лаб. исследования по ID заявки",
        "description": "Возвращает список лаб. исследований по ID заявки",
        "operationId": "getTestsByOrderId",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список заявок успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TestResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Лаб. исследования не найдены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Получить заявку по ID",
        "description": "Возвращает заявку по указанному ID",
        "operationId": "getOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Заявка не найдена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Заявка успешна найдена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "summary": "Удаление заявки",
        "description": "Удаляет заявку по указанному идентификатору",
        "operationId": "deleteOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Заявка не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "Заявка успешно удален"
          }
        }
      }
    },
    "/api/v1/orders/patient/{patientId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Получить заявки по ID пациента",
        "description": "Возвращает список заявок по ID пациента",
        "operationId": "getOrdersByPatientId",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Пациент не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Список заявок успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "TestResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID лабораторного исследования",
            "format": "int64"
          },
          "orderId": {
            "type": "integer",
            "description": "ID заявки",
            "format": "int64"
          },
          "testTypeId": {
            "type": "integer",
            "description": "ID типа исследования",
            "format": "int64"
          },
          "executionDate": {
            "type": "string",
            "description": "Дата выполнения лабораторного исследования (yyyy-MM-dd HH:mm:ss)",
            "format": "date-time"
          },
          "result": {
            "type": "string",
            "description": "Результат лабораторного исследования",
            "example": "Результат положительный"
          },
          "referenceValues": {
            "type": "string",
            "description": "Референсные значения для лабораторного исследования",
            "example": "Значение"
          },
          "status": {
            "type": "string",
            "description": "Статус лабораторного исследования",
            "example": "PENDING",
            "enum": [
              "PENDING",
              "COMPLETED"
            ]
          }
        },
        "description": "Ответ DTO для просмотра лабораторных исследований"
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Сообщение об ошибке",
            "example": "Пациентов с id-404 не найдено"
          },
          "statusCode": {
            "type": "integer",
            "description": "Код ошибки",
            "format": "int32",
            "example": 404
          }
        }
      },
...
      "OrderResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID заявки",
            "format": "int64"
          },
          "patientId": {
            "type": "integer",
            "description": "ID пациента",
            "format": "int64"
          },
          "createdDate": {
            "type": "string",
            "description": "Дата создания заявки yyyy-MM-dd HH:mm:ss",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "description": "Статус заявки",
            "example": "CANCELED",
            "enum": [
              "REGISTERED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELED"
            ]
          },
          "comment": {
            "type": "string",
            "description": "Комментарий к заявке",
            "example": "комментарий"
          }
        },
        "description": "Ответ DTO для просмотра заявки"
      },
      "OrderRequestDTO": {
        "required": [
          "patientId",
          "status"
        ],
        "type": "object",
        "properties": {
          "patientId": {
            "maximum": 9.223372036854776e+18,
            "minimum": 1,
            "type": "integer",
            "description": "ID пациента",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "description": "Статус заявки",
            "example": "CANCELED",
            "enum": [
              "REGISTERED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELED"
            ]
          },
          "comment": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "description": "Комментарий к заявке",
            "example": "комментарий"
          }
        },
        "description": "Запрос DTO для создания и обновления заявки"
      }
    }
  }
}